Refactored upgrade functions.
