+++ src/migration/template/webpack.dev.angular2.js
