+++ src/migration/template/webpack.dev.react.js
