+++ src/migration/template/webpack.dev.vue.js
