+++ src/migration/template/webpack.prod.angular2.js
