+++ src/migration/template/webpack.prod.react.js
