+++ src/migration/template/webpack.prod.vue.js
